requirements.txt
asgiref==3.8.1
Django==5.1.3
psycopg2-binary==2.9.10
sqlparse==0.5.1
typing_extensions==4.12.2
cloudinary==1.42.1
django-cloudinary-storage[video]==0.3.0
django-allauth==65.1.0
Pillow==11.0.0
environs[django]==11.0.0
weasyprint==52.5
arabic-reshaper==3.0.0
python-bidi==0.4.2
whitenoise==6.8.2
gunicorn==23.0.0
django-crispy-forms==2.3 
crispy-tailwind==1.0.3
django-widget-tweaks==1.5.0
django-extensions==3.2.3
django-debug-toolbar==4.4.6
django-tailwind[reload]==3.8.0
PyJWT[crypto]==2.9.0
google-auth==2.38.0
google-api-python-client==2.156.0
pytz==2024.2
django-ses==4.3.1
django-storages==1.14.4
requests==2.31.0
python-docx==1.1.2
django-htmx==1.21.0
XlsxWriter==3.2.0
djangorestframework==3.15.2
google-auth-oauthlib==1.2.1 
google-auth-httplib2==0.2.0 











